feat(sort-heritage-clauses): move types in their own file
