+++ rules/sort-heritage-clauses/types.ts
@@ -1,0 +1,14 @@
+import type {
+  DeprecatedCustomGroupsOption,
+  CommonOptions,
+  GroupsOptions,
+} from '../../types/common-options'
+
+export type Options = Partial<
+  {
+    customGroups: DeprecatedCustomGroupsOption
+    groups: GroupsOptions<Group>
+  } & CommonOptions
+>[]
+
+type Group = 'unknown' | string
